Handle setting site with path
Using site for both api and oauth endpoint means
that the  only  the "host part" of the site it used
for the endpoints.
So having  your api running at http://localhost/api mean that when
doing a GET to /customers was actually doing a request to http://localhost/customers
instead http://localhost/api/customers.